pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aethelum-core-lite"
version = "1.0.0"
description = "模拟树神经系统的通信框架"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Aethelum Team", email = "team@aethelum.ai"}
]
maintainers = [
    {name = "Aethelum Team", email = "team@aethelum.ai"}
]
keywords = [
    "neural-network",
    "ai",
    "moral-audit",
    "communication",
    "agent",
    "protobuf",
    "safety"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
    "requests>=2.25.0",
    "protobuf>=3.20.0",
    "toml>=0.10.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.991",
]
docs = [
    "sphinx>=4.5.0",
    "sphinx-rtd-theme>=1.0.0",
]
performance = [
    "msgpack>=1.0.0",
]
monitoring = [
    "prometheus-client>=0.19.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-jaeger>=1.20.0",
]
api = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
]
all = [
    "aethelum-core-lite[dev,docs,performance,monitoring,api]",
]

[project.urls]
Homepage = "https://github.com/aethelum/aethelum-core-lite"
Documentation = "https://aethelum-core-lite.readthedocs.io/"
Repository = "https://github.com/aethelum/aethelum-core-lite.git"
"Bug Reports" = "https://github.com/aethelum/aethelum-core-lite/issues"

[tool.setuptools]
packages = ["aethelum_core_lite"]

[tool.setuptools.package-data]
aethelum_core_lite = ["*.proto", "*.md"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "tests",
]
python_files = [
    "test_*.py",
    "*_test.py",
]